+++ source/isaaclab/test/physics/test_capabilities.py
@@ -13,7 +13,7 @@ is purely structural.
 
 from __future__ import annotations
 
-from typing import Any, ClassVar, Protocol, runtime_checkable
+from typing import ClassVar, Protocol, runtime_checkable
 
 import pytest
 
@@ -23,7 +23,6 @@ from isaaclab.physics import (
     GpuTransformBuffer,
     UsdFabric,
 )
-
 
 # ---------------------------------------------------------------------------
 # Test fixtures
@@ -66,14 +65,29 @@ class _FakeProvider(BaseSceneDataProvider):
         super().__init__()
 
     def update(self) -> None: ...
-    def get_newton_model(self): return None
-    def get_newton_state(self): return None
-    def get_usd_stage(self): return None
-    def get_metadata(self): return {}
-    def get_transforms(self): return None
-    def get_velocities(self): return None
-    def get_contacts(self): return None
-    def get_camera_transforms(self): return None
+    def get_newton_model(self):
+        return None
+
+    def get_newton_state(self):
+        return None
+
+    def get_usd_stage(self):
+        return None
+
+    def get_metadata(self):
+        return {}
+
+    def get_transforms(self):
+        return None
+
+    def get_velocities(self):
+        return None
+
+    def get_contacts(self):
+        return None
+
+    def get_camera_transforms(self):
+        return None
 
 
 class _ConsumerWantingGpu:
@@ -85,9 +99,7 @@ class _ConsumerWantingUsd:
 
 
 class _ConsumerWantingEither:
-    required_one_of: ClassVar[tuple[tuple[type, ...], ...]] = (
-        (UsdFabric, GpuTransformBuffer),
-    )
+    required_one_of: ClassVar[tuple[tuple[type, ...], ...]] = ((UsdFabric, GpuTransformBuffer),)
 
 
 class _ConsumerWantingBoth: